fix chroma tools

+++ llamaindex/vector_store/chroma_clean.py
@@ -1,9 +1,15 @@
 import os, sys, chromadb
 
-path = os.environ.get("CHROMA_DB_DIR", "chroma")
-db = chromadb.PersistentClient(path=path)
-collection = len(sys.argv) == 2 and sys.argv[1] or None
-if collection is None:
-    print("provide the collection name")
+base_dir = os.getenv("CHROMA_BASE_DIR", "chroma")
+db_dir = len(sys.argv) >= 2 and sys.argv[1] or None
+
+if db_dir:
+    db = chromadb.PersistentClient(path=os.path.join(base_dir, db_dir))
+    collection = len(sys.argv) == 3 and sys.argv[2] or None
+    if collection is None:
+        print("provide the collection name")
+    else:
+        db.delete_collection(collection)
 else:
-    db.delete_collection(collection)
+    for subpath in os.listdir(base_dir):
+        print(subpath)
